+++ graphdatascience/query_runner/neo4j_query_runner.py
@@ -5,7 +5,7 @@ import re
 import time
 import warnings
 from concurrent.futures import Future, ThreadPoolExecutor, wait
-from typing import Any, Dict, List, Optional, Tuple, Union
+from typing import Any, Dict, List, NoReturn, Optional, Tuple, Union
 from uuid import uuid4
 
 import neo4j
@@ -226,14 +226,19 @@ class Neo4jQueryRunner(QueryRunner):
             self._logger.info(notification)
 
     def _log(self, job_id: str, future: "Future[Any]", database: Optional[str] = None) -> None:
-        pbar = None
+        pbar: Optional[tqdm[NoReturn]] = None
         warn_if_failure = True
 
         while wait([future], timeout=self._LOG_POLLING_INTERVAL).not_done:
             try:
                 tier = "beta." if self._server_version < ServerVersion(2, 5, 0) else ""
+                # we only retrieve the progress of the root task
                 progress = self.run_cypher(
-                    f"CALL gds.{tier}listProgress('{job_id}') YIELD taskName, progress", database=database
+                    f"CALL gds.{tier}listProgress('{job_id}')"
+                    + " YIELD taskName, progress"
+                    + " RETURN taskName, progress"
+                    + " LIMIT 1",
+                    database=database,
                 )
             except Exception as e:
                 # Do nothing if the procedure either:
@@ -248,17 +253,19 @@ class Neo4jQueryRunner(QueryRunner):
                     continue
 
             progress_percent = progress["progress"][0]
-            if not progress_percent == "n/a":
-                task_name = progress["taskName"][0].split("|--")[-1][1:]
-                pbar = pbar or tqdm(total=100, unit="%", desc=task_name)
-            else:
+            if progress_percent == "n/a":
                 return
+
+            root_task_name = progress["taskName"][0].split("|--")[-1][1:]
+            if not pbar:
+                pbar = tqdm(total=100, unit="%", desc=root_task_name, maxinterval=self._LOG_POLLING_INTERVAL)
 
             parsed_progress = float(progress_percent[:-1])
             pbar.update(parsed_progress - pbar.n)
 
         if pbar:
             pbar.update(100 - pbar.n)
+            pbar.refresh()
 
     def set_database(self, database: str) -> None:
         self._database = database