+++ changelog.md
@@ -20,6 +20,7 @@
 * Fixed an issue where source and target IDs of relationships in heterogeneous OGBL graphs were not parsed correctly.
 * Fixed an issue where configuration parameters such as `aggregation` were ignored by `gds.graph.toUndirected`.
 * Fixed an issue where the `database` given for the `GraphDataScience` construction was not used for metadata retrieval, causing an exception to be raised if the default "neo4j" database was missing.
+* Fixed an issue where progress bars would not always complete.
 
 
 ## Improvements